Update website spacing
Use 1 as the spacing unit size, instead of 1.25.
Bleed the footer background color to the full page width.
Make the team photos grayscale.
Make logo in header larger.
Update the product images and make them PNGs.
Use a Meetup link for the Learn More CTA and a mailto for Contat Us.
Overlap the feature rows on the home page.
Add sx prop to Logo and align it vertically.
Remove GIF nonsense.